nftw: add last missing flag
With the implementation of FTW_ACTIONRETVAL, it is feature complete.
However untested, and there is at least 1 rough part: when siblings should
be skipped, when not using FTW_DEPTH, and a folder is being processed.
That supposed to work, but there is no way that I will understand that
in a week. That will need rework.